requirements.txt
Django==4.0.4
django-cleanup==6.0.0
django-cors-headers==3.11.0
django-jazzmin==2.5.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
Pillow==9.1.0
psycopg2-binary==2.9.3
pytz==2022.1
